refac: TablsComps에 각 데이터별 table field를 외부 infofiled.js로 추출
