+++ test-web/src/components/DataDetailPage/TablesComps/LabTable.js
@@ -7,7 +7,7 @@ import {
   TableHead,
   TableRow,
 } from '@mui/material';
-
+import { labField, labDBFieldToSemanticWord } from '../constants/infofield';
 const LabTable = ({
   edited,
   labInput,
@@ -28,15 +28,12 @@ const LabTable = ({
           <TableRow key={'labData-explanation'}>
             <TableCell key={'labData-exp-col'}>{}</TableCell>
             <TableCell key={'labData-exp-col0'}>원육</TableCell>
-            {Array.from(
-              { length: len},
-              (_, arr_idx) => (
-                <TableCell key={'labData-exp-col' + (arr_idx + 1)}>
-                  {' '}
-                  {processed_data_seq[arr_idx + 1]}차{' '}
-                </TableCell>
-              )
-            )}
+            {Array.from({ length: len }, (_, arr_idx) => (
+              <TableCell key={'labData-exp-col' + (arr_idx + 1)}>
+                {' '}
+                {processed_data_seq[arr_idx + 1]}차{' '}
+              </TableCell>
+            ))}
           </TableRow>
         </TableHead>
 
@@ -69,33 +66,30 @@ const LabTable = ({
                 </TableCell>
                 {
                   // 실험실 및 가열육 추가 데이터 수정
-                  Array.from(
-                    { length: len },
-                    (_, arr_idx) => (
-                      <TableCell key={'lab-' + arr_idx + '-col' + arr_idx}>
-                        {edited ? (
-                          <input
-                            key={'lab-' + arr_idx + '-input'}
-                            style={{ width: '100px', height: '23px' }}
-                            name={f}
-                            value={labInput[arr_idx + 1]?.[f]}
-                            placeholder={
-                              lab_data[arr_idx + 1] === null
-                                ? '0.0'
-                                : lab_data[arr_idx]?.[f]
-                            }
-                            onChange={(e) => {
-                              handleInputChange(e, 3, arr_idx + 1);
-                            }}
-                          />
-                        ) : labInput[arr_idx + 1]?.[f] ? (
-                          labInput[arr_idx + 1]?.[f]
-                        ) : (
-                          ''
-                        )}
-                      </TableCell>
-                    )
-                  )
+                  Array.from({ length: len }, (_, arr_idx) => (
+                    <TableCell key={'lab-' + arr_idx + '-col' + arr_idx}>
+                      {edited ? (
+                        <input
+                          key={'lab-' + arr_idx + '-input'}
+                          style={{ width: '100px', height: '23px' }}
+                          name={f}
+                          value={labInput[arr_idx + 1]?.[f]}
+                          placeholder={
+                            lab_data[arr_idx + 1] === null
+                              ? '0.0'
+                              : lab_data[arr_idx]?.[f]
+                          }
+                          onChange={(e) => {
+                            handleInputChange(e, 3, arr_idx + 1);
+                          }}
+                        />
+                      ) : labInput[arr_idx + 1]?.[f] ? (
+                        labInput[arr_idx + 1]?.[f]
+                      ) : (
+                        ''
+                      )}
+                    </TableCell>
+                  ))
                 }
               </TableRow>
             );
@@ -107,39 +101,3 @@ const LabTable = ({
 };
 
 export default LabTable;
-
-const labField = [
-  'L',
-  'a',
-  'b',
-  'DL',
-  'CL',
-  'RW',
-  'ph',
-  'WBSF',
-  'cardepsin_activity',
-  'MFI',
-  'sourness',
-  'bitterness',
-  'umami',
-  'richness',
-  'Collagen',
-];
-const labDBFieldToSemanticWord = {
-  L: '명도',
-  a: '적색도',
-  b: '황색도',
-  DL: '육즙감량',
-  CL: '가열감량',
-  RW: '압착감량',
-  ph: 'pH',
-  WBSF: '전단가',
-  cardepsin_activity: '카뎁신활성도',
-  MFI: '근소편화지수',
-  sourness: '신맛',
-  bitterness: '진한맛',
-  umami: '감칠맛',
-  richness: '후미',
-  Collagen: '콜라겐',
-};
-
