+++ test-web/src/components/DataDetailPage/style/meatimgscardstaticstyle.js
@@ -1,0 +1,72 @@
+const navy = '#0F3659';
+
+const style = {
+    imgTitleContainer: {
+      display: 'flex',
+      justifyContent: 'space-between',
+      alignItems: 'center',
+    },
+    imgTitleWrapper: {
+      color: '#002984',
+      height: '40px',
+      fontSize: '18px',
+      fontWeight: '800',
+    },
+    imgContainer: {
+      height: '350px',
+      width: '100%',
+      borderRadius: '10px',
+    },
+    imgWrapper: {
+      height: '350px',
+      width: '400px',
+      objectFit: 'contain',
+    },
+    imgNotExistWrapper: {
+      height: '350px',
+      width: '400px',
+      display: 'flex',
+      justifyContent: 'center',
+      alignItems: 'center',
+    },
+    paginationBtnWrapper: {
+      display: 'flex',
+      justifyContent: 'center',
+      alignItems: 'center',
+    },
+    paginationLeftBtn: {
+      height: '35px',
+      width: '35px',
+      borderRadius: '10px',
+      padding: '0',
+      border: '1px solid black',
+    },
+    paginationNavBtnWrapper: {
+      display: 'flex',
+      justifyContent: 'center',
+      margin: '0px 5px',
+    },
+    paginationNavCurrDiv: {
+      height: 'fit-content',
+      width: 'fit-content',
+      padding: '10px',
+      borderRadius: '5px',
+      color: navy,
+    },
+    paginationNavNotCurrDiv: {
+      height: '100%',
+      width: 'fit-content',
+      borderRadius: '5px',
+      padding: '10px',
+      color: '#b0bec5',
+    },
+    paginationRightBtn: {
+      height: '35px',
+      width: '35px',
+      borderRadius: '10px',
+      padding: '0',
+      border: '1px solid black',
+    },
+  };
+  
+export default style
