+++ test-web/src/components/DataDetailPage/style/meatimgscardstyle.js
@@ -1,0 +1,94 @@
+const navy = '#0F3659';
+const style = {
+    imgTitleContainer: {
+      display: 'flex',
+      justifyContent: 'space-between',
+      alignItems: 'center',
+    },
+    imgTitleWrapper: {
+      color: '#002984',
+      height: '40px',
+      fontSize: '18px',
+      fontWeight: '800',
+    },
+    imgContainer: {
+      height: '350px',
+      width: '100%',
+      borderRadius: '10px',
+      display: 'flex',
+      justifyContent: 'center',
+      alignItems: 'center',
+    },
+    imgWrapper: {
+      maxHeight: '350px',
+      maxWidth: '100%',
+      objectFit: 'contain',
+    },
+    imgNotExistWrapper: {
+      height: '350px',
+      width: '400px',
+      display: 'flex',
+      justifyContent: 'center',
+      alignItems: 'center',
+    },
+    imgUploadWrapper: {
+      height: '40px',
+      width: '160px',
+      border: `1px solid ${navy}`,
+      borderRadius: '5px',
+    },
+    imgUploadTextWrapper: {
+      color: `${navy}`,
+      fontSize: '16px',
+      marginRight: '5px',
+    },
+    paginationBtnWrapper: {
+      display: 'flex',
+      justifyContent: 'center',
+      alignItems: 'center',
+      flexWrap: 'wrap',
+      gap: '5px',
+    },
+    paginationLeftBtn: {
+      height: '35px',
+      width: '35px',
+      borderRadius: '10px',
+      padding: '0',
+      border: '1px solid black',
+      marginBottom: '5px',
+    },
+    paginationNavBtnWrapper: {
+      display: 'flex',
+      justifyContent: 'center',
+      flexWrap: 'wrap',
+      gap: '5px',
+    },
+    paginationNavCurrDiv: {
+      height: '35px',
+      width: '35px',
+      display: 'flex',
+      justifyContent: 'center',
+      alignItems: 'center',
+      borderRadius: '5px',
+      color: navy,
+    },
+    paginationNavNotCurrDiv: {
+      height: '35px',
+      width: '35px',
+      display: 'flex',
+      justifyContent: 'center',
+      alignItems: 'center',
+      borderRadius: '5px',
+      color: '#b0bec5',
+    },
+    paginationRightBtn: {
+      height: '35px',
+      width: '35px',
+      borderRadius: '10px',
+      padding: '0',
+      border: '1px solid black',
+      marginBottom: '5px',
+    },
+  };
+  
+export default style
